Use a safe access method for key lookups on call
Previous method would fail if the VCF info did not contain all the necessary fields. This failure would be in the form of a KeyError which would cause the transitions to fail resulting in an infinite loop in the sample pipeline handler which would always fail and log STS error with no additional info.