Remove code requiring AD, DP, GQ, GT, & PL in VCF info to load result
Previously, if one or more of these fields was absent from the INFO section the result would not be loaded. Now, if these fields are present, the data is added to the result but if it is missing, then the result is created with null values for these fields. We cannot guarantee that everyone will have or want these fields in their VCF files so the pipeline now handles them when they are there but does not make any assumptions about their existence.