Lookup by batch/sample name instead of pk when sample result counts
This fixes an issue exposed by the Travis CI tests. We cannot guarantee the order in which the samples are created so we therefore have no guarantee that the pk, result count tuples will be correct. Using sample/batch name combos is the only sure way to make sure the result counts are matched to the correct sample.