Use a single template file for variant/result templated data
I tried to put these into views within cell.coffee but they are used in the details modal as well so it was easier to just go back to using the original varify method of using a bunch of template methods in a single templates file. This can be changed later but this way allows for the easiest reuse in the details modal.