Make assessment.assessment_category a required column
The addition of the "Other" assessment(via a data migration in this change set) will account for any existing assessments that have a null assessment_category field. This change will allow the UI to make category a selectable field for all pathogenicities instead of the limited subset as it is now.