Force user to select a sample
This has the side effect of blocking all context changes until the sample is selected but that is necessary given the way required concepts are implemented currently in Cilantro. There is a notification made to the user when they try to modify the context with no sample selected so it should be clear enough that no one is confused.