Fix bad data lookups on PyVCF call objects
Originally, all these lookups were done using explicit indexing of the form call['GT'] for example. This caused AttributeErrors when the key was not present. The move was made to getattr(call, 'GQ', None) to try to safely access those attributes. This unknowingly resulted in all None values being returned because the call object does not hold this data directly, it is all on its data property which is what is read in explicit indexing because PyVCF overrides the __getitem__ method(https://github.com/cbmi/PyVCF/blob/master/vcf/model.py#L90-L92). To avoid AttributeErrors, we now use a safe lookup on call.data instead of call which will read the data in a safe manner unlike the explicit indexing case because PyVCF always assumes the key is valid when calling getattr.