Require one or more columns to be selected before exporting
The error message is now hidden when all conditions are satisfied and
the export is about to begin. The reason we don't hide the message when
the user changes column or type or page range is that the same error
container is used for all 3 of the error cases. Currently, it seems like
more work to monitor all the different controls for changes just to
show/hide the error. Easier to do all the checks when the user clicks
the export button. This should be simplified when Cilantro refactors the
export related code.