Remove selectors hash in favor of ui and rebinding
Since many of the ui elements we need to address are not available in
the templates, they will not be bound during the normal pre-render
bindUIElements call. As a result, we rebind the UI elements after all
the regions are shown or in the onRender method before interacting with
the ui object.

Signed-off-by: Don Naegely <naegelyd@gmail.com>