Update templates module to pass array to template registry
This takes advantage of the custom tpl plugin Cilantro provides which
converts the module path to a suitable template name. This combines the
path of the project, renamed from 'project' to 'varify' and the path
with the relative tokens or templates/ directory. For example, a path
of '../../templates/variant/summary.html' resolves to
'varify/variant/summary'.

The phenotype modal template has been renamed to matched the referenced
name in the view.

Signed-off-by: Byron Ruth <b@devel.io>